pyproject.toml
[tool.poetry]
name = "coffeebase-web"
version = "0.1.0"
description = ""
authors = ["Victor <A01643020@tec.mx>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^3.0.2"
python-dotenv = "^1.0.1"
schedule = "^1.2.1"
urllib3 = "^2.2.1"
firebase-admin = "^6.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
